Reorganiziation + Extra Util
