+++ PathPlanner/src/main/java/com/millburnx/purePursuit/Utils/Intersection.kt
@@ -1,4 +1,5 @@
-package com.millburnx.purePursuit
+package com.millburnx.purePursuit.Utils
+
 class Intersection(val point: Point, val line: Line) {
     override fun toString(): String {
         return "Intersection(point=$point, line=$line)"
